crewagent-campaignplanner/requirements.txt
crewai
crewai_tools
pydantic
python-dotenv